[Extraction] prep for go.mod & go.work FilePath ReplaceDirective work
