+++ internal/go_repository_tools_srcs.bzl
@@ -8,7 +8,9 @@ GO_REPOSITORY_TOOLS_SRCS = [
     Label("//cmd/autogazelle:client_unix.go"),
     Label("//cmd/autogazelle:server_unix.go"),
     Label("//cmd/fetch_repo:BUILD.bazel"),
+    Label("//cmd/fetch_repo:copy_tree.go"),
     Label("//cmd/fetch_repo:fetch_repo.go"),
+    Label("//cmd/fetch_repo:go_mod_download.go"),
     Label("//cmd/fetch_repo:module.go"),
     Label("//cmd/fetch_repo:vcs.go"),
     Label("//cmd/gazelle:BUILD.bazel"),
